split plugin tester
