+++ tests/playwright/sanity/plugin-tester-container.test.ts-snapshots/frontPage-linux.png
