+++ tests/playwright/sanity/plugin-tester-section.test.ts-snapshots/.gitignore
@@ -1,0 +1,3 @@
+*.*
+!*-linux.*
+!.gitignore
